requirements.txt
absl-py==2.3.1
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
ansi2html==1.8.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
async-timeout==5.0.1
attrs==22.2.0
audioread==3.0.1
Babel==2.12.1
backcall==0.2.0
backports-datetime-fromisoformat==2.0.3
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.11.2
bleach==6.0.0
braceexpand==0.1.7
Brotli==1.1.0
brotlipy==0.7.0
category-encoders==2.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.2.1
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.0.7
cryptography==39.0.1
cssselect2==0.8.0
cuda-bindings==12.9.0
cuda-python==12.9.0
cycler==0.11.0
Cython==0.29.33
cytoolz==1.0.1
dash==2.9.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasets==3.6.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
deprecation==2.1.0
dill==0.3.8
Distance==0.1.3
docopt==0.6.2
editdistance==0.8.1
einops==0.8.1
entrypoints==0.4
executing==1.2.0
fastapi==0.116.1
fastjsonschema==2.16.3
fiddle==0.3.0
filelock==3.9.0
Flask==2.2.3
flatbuffers==23.3.3
flit_core==3.6.0
fonttools==4.39.2
frozenlist==1.7.0
fsspec==2024.12.0
future==1.0.0
g2p-en==2.1.0
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.20.1
greenlet==3.2.3
grpcio==1.73.1
h11==0.16.0
h5py==3.14.0
huggingface-hub==0.33.4
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
imbalanced-learn==0.10.1
importlib-metadata==6.0.0
importlib-resources==5.12.0
inflect==7.5.0
intervaltree==3.1.0
ipykernel==6.15.0
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jiwer==3.1.0
joblib==1.2.0
json5==0.9.5
jsonschema==4.17.3
jupyter_client==8.0.3
jupyter-contrib-core==0.4.0
jupyter-contrib-nbextensions==0.7.0
jupyter_core==5.3.0
jupyter-dash==0.4.2
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.6.1
jupyter-server==1.23.6
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.20.0
jupyterlab-widgets==3.0.5
kaldi-python-io==1.2.2
kaldiio==2.18.1
kaleido==0.2.1
kiwisolver==1.4.4
langdetect==1.0.9
lazy_loader==0.4
Levenshtein==0.27.1
lhotse==1.30.3
libcst==1.8.2
librosa==0.11.0
lightgbm==3.3.5
lightning==2.4.0
lightning-utilities==0.14.3
lilcom==1.8.1
llvmlite==0.44.0
loguru==0.7.3
lxml==4.9.1
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.1
marshmallow==4.0.0
matplotlib==3.10.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapy==1.1.6
mistune==2.0.5
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
more-itertools==10.7.0
mplcursors==0.6
mpmath==1.2.1
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
nb-conda-kernels==2.3.1
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nemo-toolkit==2.3.2
nest-asyncio==1.5.6
networkx==2.8.4
nltk==3.9.1
notebook==6.5.3
notebook_shim==0.2.2
num2words==0.5.14
numba==0.61.0
numpy==1.26.4
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.16.3
opencv-python==4.7.0.72
optuna==4.4.0
orjson==3.8.7
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
peft==0.16.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.0.1
pkgutil_resolve_name==1.3.10
plac==1.4.5
platformdirs==2.5.2
plotly==5.13.1
plotly-resampler==0.8.3.2
pmdarima==2.0.3
pooch==1.8.2
prometheus-client==0.16.0
prompt-toolkit==3.0.38
propcache==0.3.2
protobuf==4.24.4
psutil==5.9.0
pure-eval==0.2.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyarrow==20.0.0
pybind11==3.0.0
pycaret==3.0.0
pyclipper==1.3.0.post6
pycocotools==2.0.6
pycparser==2.21
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pydyf==0.11.0
Pygments==2.14.0
pyloudnorm==0.1.1
pyod==1.0.8
pyOpenSSL==23.0.0
pyparsing==3.0.9
pypdfium2==4.30.1
pyphen==0.17.2
pyreadline3==3.4.1
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-doctr==1.0.0
python-multipart==0.0.20
pytorch-lightning==2.0.4
pytz==2022.7.1
pywin32==305.1
pywinpty==2.0.10
PyYAML==6.0
pyzmq==23.2.0
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.4
resampy==0.4.3
retrying==1.3.4
rich==14.0.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
sacremoses==0.1.1
safetensors==0.5.3
schemdraw==0.15
scikit-learn==1.2.2
scikit-plot==0.3.7
scipy==1.10.1
seaborn==0.12.2
segment-anything==1.0
Send2Trash==1.8.0
sentencepiece==0.2.0
sentry-sdk==2.32.0
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.16.0
sklearn==0.0.post1
sktime==0.16.1
smmap==5.0.2
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.3.2.post1
sox==1.5.0
soxr==0.5.0.post1
SQLAlchemy==2.0.41
stack-data==0.6.2
starlette==0.47.1
statsmodels==0.13.5
sympy==1.14.0
tabulate==0.9.0
tbats==1.1.2
tenacity==8.2.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminado==0.15.0
text-unidecode==1.3
texterrors==0.5.1
threadpoolctl==3.1.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.2
tomli==2.0.1
toolz==1.0.0
torch==2.7.1+cu126
torchaudio==2.7.1+cu126
torchmetrics==0.11.4
torchvision==0.22.1+cu126
torchviz==0.0.2
tornado==6.2
tqdm==4.67.1
trace-updater==0.0.9
traitlets==5.9.0
transformers==4.53.2
typeguard==4.4.4
typer==0.16.0
typing_extensions==4.14.1
typing-inspection==0.4.1
urllib3==1.26.14
uvicorn==0.35.0
validators==0.35.0
wandb==0.21.0
wcwidth==0.2.6
weasyprint==65.1
webdataset==1.0.2
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
wget==3.2
wheel==0.38.4
widgetsnbextension==4.0.5
win-inet-pton==1.1.0
win32_setctime==1.2.0
wincertstore==0.2
wrapt==1.15.0
xxhash==3.2.0
yarl==1.20.1
yellowbrick==1.5
zipp==3.15.0
zopfli==0.2.3.post1
